[main] download outer awb file correctly #19

+++ app/src/ts/renderer/mishiro-home.ts
@@ -164,7 +164,7 @@ export default class extends Vue {
 
           this.current = 0
           this.text = ''
-          if (suffix !== '.acb') {
+          if (suffix !== '.acb' && suffix !== '.awb') {
             if (fs.existsSync(filepath)) {
               fs.removeSync(filepath)
               this.event.$emit('completeTask', name + suffix)
